requirements.txt
aiofiles==0.5.0
appdirs==1.4.4
attrs==19.3.0
black==19.10b0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask-WTF==0.14.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.7.7
httptools==0.1.1
httpx==0.11.1
hyperframe==5.2.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==2.21.0
motor==2.1.0
multidict==4.7.6
pathspec==0.8.0
pymongo==3.10.1
python-dateutil==2.8.1
regex==2020.6.8
rfc3986==1.4.0
sanic==20.6.3
sanic-base-extension==0.1.1
sanic-mongodb-extension==0.4.0
six==1.15.0
sniffio==1.1.0
toml==0.10.1
typed-ast==1.4.1
ujson==3.0.0
umongo==2.2.0
urllib3==1.25.9
uvloop==0.14.0
websockets==8.1
